+++ src/preload/commands/restrict/index.ts
@@ -64,7 +64,7 @@ export default class Restrict extends Command {
 
         await server.save();
 
-        await this.respond(msg, lang.restrict_success.replace('%channels', (allChannels ? lang.restrict_allcmnds : channels.map(channelID => `<#${channelID}>`).join(", "))).replace('%cmds', (commands.length > 0) ? (allCommands ? lang.restrict_allchns : commands.join("\n")) : lang.restrict_none));
+        await this.respond(msg, lang.restrict_success.replace('%channels', (allChannels ? lang.restrict_allchns : channels.map(channelID => `<#${channelID}>`).join(", "))).replace('%cmds', (commands.length > 0) ? (allCommands ? lang.restrict_allcmnds : commands.join("\n")) : lang.restrict_none));
 
         return true;
     }